latest

+++ py/testdir_release/runner.sh
@@ -73,7 +73,11 @@ ls -lt ./h2o-nodes.json
 # We don't want to hang waiting for the cloud to terminate.
 ../testdir_single_jvm/n0.doit test_shutdown.py
 
-kill $CLOUD_PID
+if ps -p $CLOUD_PID > /dev/null
+then
+    echo "$CLOUD_PID is still running after shutdown. Will kill"
+    kill $CLOUD_PID
+fi
 ps aux | grep four_hour_cloud
 
 # test_c2_rel has about 11 subtests inside it, that will be tracked individually by jenkins
